Reduce CopilotHeroBox to panel helpers and rename to copilotPanel
- Drop the unused inline hero-box component; keep only openCopilotPanel/submitPromptToCopilot
- Rename the module to copilotPanel.ts and update its importer
